Finished workplan validation rules such as spam or public/private update
Dashboard additionaly finished

+++ src/main/java/acme/entities/dashboard/Dashboard.java
@@ -45,6 +45,6 @@ public class Dashboard implements Serializable {
 	Double 	dvtExecutionPeriodWorkplans;
 	
 	Integer totalNumberOfWorkplans;
-	Integer totalNumberOfPublishedWorkplans;
-	Integer totalNumberOfNonPublishedWorkplans;
+	Long totalNumberOfPublishedWorkplans;
+	Long totalNumberOfNonPublishedWorkplans;
 }
